MLJ.jl

MLJ.jl

Documentation for MLJ.jl

Implementing the MLJ interface for a learning algorithm

Overview

The Model API

MLJ.jl

Functions

Index

MLJ Internals

The machine interface, simplified

Functions

MLJ.@curve — Macro.

@curve

The code,

@curve var range code

evaluates code, replacing appearances of var therein with each value in range. The range and corresponding evaluations are returned as a tuple of arrays. For example,

@curve  x 1:3 (x^2 + 1)

evaluates to

([1,2,3], [2, 5, 10])

This is convenient for plotting functions using, eg, the Plots package:

plot(@curve x 1:3 (x^2 + 1))

A macro @pcurve parallelizes the same behaviour. A two-variable implementation is also available, operating as in the following example:

julia> @curve x [1,2,3] y [7,8] (x + y)
([1,2,3],[7 8],[8.0 9.0; 9.0 10.0; 10.0 11.0])

julia> ans[3]
3×2 Array{Float64,2}:
  8.0   9.0
  9.0  10.0
 10.0  11.0

N.B. The second range is returned as a row vector for consistency with the output matrix. This is also helpful when plotting, as in:

julia> u1, u2, A = @curve x range(0, stop=1, length=100) α [1,2,3] x^α
julia> u2 = map(u2) do α "α = "*string(α) end
julia> plot(u1, A, label=u2)

which generates three superimposed plots - of the functions x, x^2 and x^3 - each labels with the exponents α = 1, 2, 3 in the legend.

source

MLJ.EnsembleModel — Method.

EnsembleModel(atom=nothing, weights=Float64[], bagging_fraction=0.8, rng_seed=0, n=100, parallel=true)

Create a model for training an ensemble of n learners, with optional bagging, each with associated model atom. Ensembling is useful if fit!(machine(atom, data...)) does not create identical models on repeated calls (ie, is a stochastic model, such as a decision tree with randomized node selection criteria), or if bagging_fraction is set to a value not equal to 1.0 (or both). The constructor fails if no atom is specified.

Predictions are weighted according to the vector weights (to allow for external optimization) except in the case that atom is a Deterministic classifier. Uniform weights are used if weight has zero length.

The ensemble model is Deterministic or Probabilistic, according to the corresponding supertype of atom. In the case of classifiers (targetscitype(atom) <: Union{Multiclass,FiniteOrderedFactor}), the predictions are majority votes, and for regressors (targetscitype(atom)<: Continuous) they are ordinary averages. Probabilistic predictions are obtained by averaging the atomic probability distribution functions; in particular, for regressors, the ensemble prediction on each input pattern has the type MixtureModel{VF,VS,D} from the Distributions.jl package, where D is the type of predicted distribution for atom.

source

MLJ.evaluate! — Method.

evaluate!(mach, resampling=CV(), measure=nothing, operation=predict, verbosity=1)

Estimate the performance of a machine mach using the specified resampling (defaulting to 6-fold cross-validation) and measure. In general this mutating operation preserves only mach.args (the data stored in the machine).

If no measure is specified, then default_measure(mach.model) is used, unless this default is nothing and an error is thrown.

source

MLJ.iterator — Method.

iterator(model::Model, param_iterators::Params)

Iterator over all models of type typeof(model) defined by param_iterators.

Each name in the nested :name => value pairs of param_iterators should be the name of a (possibly nested) field of model; and each element of flat_values(param_iterators) (the corresponding final values) is an iterator over values of one of those fields.

See also iterator and params.

source

MLJ.learning_curve — Method.

learning_curve(model, X, ys...; resolution=30, resampling=Holdout(), measure=rms, operation=pr, param_range=nothing)

Returns (u, v) where u is a vector of hyperparameter values, and v the corresponding performance estimates.

X, y = datanow()
atom = RidgeRegressor()
model = EnsembleModel(atom=atom)
r = range(atom, :lambda, lower=0.1, upper=100, scale=:log10)
param_range = Params(:atom => Params(:lambda => r))
u, v = MLJ.learning_curve(model, X, y; param_range = param_range) 

source

MLJ.models — Method.

models()

List the names of all MLJ models, loaded or registered, as a dictionary indexed on package name.

source

MLJ.rmsp — Method.

Root mean squared percentage loss

source

MLJ.strange — Method.

strange(model, :hyper; kwargs...)

Returns the pair :hyper => range(model, :hyper; kwargs...); "strange" is short for "set to range".

See also: range

source

MLJBase.info — Function.

info(model, pkg=nothing)

Return the dictionary of metadata associated with model::String. If more than one package implements model then pkg::String will need to be specified.

source

MLJ.SCALE — Type.

Scale = SCALE()

Object for dispatching on scales and functions when generating parameter ranges. We require different behaviour for scales and functions:

 transform(Scale, scale(:log), 100) = 2
 inverse_transform(Scale, scale(:log), 2) = 100

but transform(Scale, scale(log), 100) = 100 # identity inverse_transform(Scale, scale(log), 100) = 2

See also: strange

source

MLJ.SimpleDeterministicCompositeModel — Type.

SimpleDeterministicCompositeModel(;regressor=ConstantRegressor(), 
                          transformer=FeatureSelector())

Construct a composite model consisting of a transformer (Unsupervised model) followed by a Deterministic model. Mainly intended for internal testing .

source

Base.copy — Function.

copy(params::Params, values=nothing)

Return a copy of params with new values. That is, flat_values(copy(params, values)) == values is true, while the first element of each nested pair (parameter name) is unchanged.

If values is not specified a deep copy is returned.

source

Base.merge! — Method.

merge!(tape1, tape2)

Incrementally appends to tape1 all elements in tape2, excluding any element previously added (or any element of tape1 in its initial state).

source

Base.range — Method.

r = range(model, :hyper; values=nothing)

Defines a NominalRange object for a field hyper of model. Note that r is not directly iterable but iterator(r) iterates over values.

r = range(model, :hyper; upper=nothing, lower=nothing, scale=:linear)

Defines a NumericRange object for a field hyper of model. Note that r is not directly iteratable but iterator(r, n) iterates over n values between lower and upper values, according to the specified scale. The supported scales are :linear, :log, :log10, :log2. Values for Integer types are rounded (with duplicate values removed, resulting in possibly less than n values).

Alternatively, if a function f is provided as scale, then iterator(r, n) iterates over the values [f(x1), f(x2), ... , f(xn)], where x1, x2, ..., xn are linearly spaced between lower and upper.

See also: strange, iterator

source

MLJ.flat_keys — Method.

 flat_keys(params::Params)

Use dot-concatentation to express each key in key-value pair of params in string form.

Example

julia> flat_keys(Params(:A => Params(:x => 2, :y => 3), :B)))
["A.x", "A.y", "B"]

source

MLJ.get_type — Method.

get_type(T, field::Symbol)

Returns the type of the field field of DataType T. Not a type-stable function.

source

MLJ.unwind — Method.

unwind(iterators...)

Represent all possible combinations of values generated by iterators as rows of a matrix A. In more detail, A has one column for each iterator in iterators and one row for each distinct possible combination of values taken on by the iterators. Elements in the first column cycle fastest, those in the last clolumn slowest.

Example

julia> iterators = ([1, 2], ["a","b"], ["x", "y", "z"]);
julia> MLJ.unwind(iterators...)
12×3 Array{Any,2}:
 1  "a"  "x"
 2  "a"  "x"
 1  "b"  "x"
 2  "b"  "x"
 1  "a"  "y"
 2  "a"  "y"
 1  "b"  "y"
 2  "b"  "y"
 1  "a"  "z"
 2  "a"  "z"
 1  "b"  "z"
 2  "b"  "z"

source

Index

MLJ.SCALE

MLJ.SimpleDeterministicCompositeModel

Base.copy

Base.merge!

Base.range

MLJ.EnsembleModel

MLJ.evaluate!

MLJ.flat_keys

MLJ.get_type

MLJ.iterator

MLJ.learning_curve

MLJ.models

MLJ.rmsp

MLJ.strange

MLJ.unwind

MLJBase.info

MLJ.@curve