Tuning models

Tuning models

API

Base.range — Function.

r = range(model, :hyper; values=nothing)

Defines a NominalRange object for a field hyper of model, assuming the field is a not a subtype of Real. Note that r is not directly iterable but iterator(r) iterates over values.

A nested hyperparameter is specified using dot notation. For example, :(atom.max_depth) specifies the :max_depth hyperparameter of the hyperparameter :atom of model.

r = range(model, :hyper; upper=nothing, lower=nothing, scale=:linear)

Defines a NumericRange object for a Real field hyper of model. Note that r is not directly iteratable but iterator(r, n) iterates over n values between lower and upper values, according to the specified scale. The supported scales are :linear, :log, :log10, :log2. Values for Integer types are rounded (with duplicate values removed, resulting in possibly less than n values).

Alternatively, if a function f is provided as scale, then iterator(r, n) iterates over the values [f(x1), f(x2), ... , f(xn)], where x1, x2, ..., xn are linearly spaced between lower and upper.

See also: iterator.

source

MLJ.TunedModel — Function.

tuned_model = TunedModel(; model=nothing,
                         tuning=Grid(),
                         resampling=Holdout(),
                         measure=nothing,
                         weights=nothing,
                         operation=predict,
                         ranges=ParamRange[],
                         full_report=true)

Construct a model wrapper for hyperparameter optimization of a supervised learner.

Calling fit!(mach) on a machine mach=machine(tuned_model, X, y) or mach=machine(tuned_model, task) will:

Instigate a search, over clones of model with the hyperparameter mutations specified by ranges, for a model optimizing the specified measure, using performance evaluations carried out using the specified tuning strategy and resampling strategy.

Fit a machine, mach_optimal = fitted_params(mach).best_model, wrapping the optimal model object in all the provided data X, y (or in task). Calling predict(mach, Xnew) then returns predictions on Xnew of the machine mach_optimal.

If a custom measure measure is used, and the measure is a score, rather than a loss, be sure to check that MLJ.orientation(measure) == :score to ensure maximization of the measure, rather than minimization. Overide an incorrect value with MLJ.orientation(::typeof(measure)) = :score.

If measure supports sample weights (MLJ.supports_weights(measure) == true) then these can be passed to the measure as weights.

In the case of two-parameter tuning, a Plots.jl plot of performance estimates is returned by plot(mach) or heatmap(mach).

source