Working with Tasks

Working with Tasks

In MLJ a task is a synthesis of three elements: data, an interpretation of that data, and a learning objective. Once one has a task one is ready to choose learning models.

Scientific types and the interpretation of data

Generally the columns of a table, such as a DataFrame, represents real quantities. However, the nature of a quantity is not always clear from the representation. For example, we might count phone calls using the UInt32 type but also use UInt32 to represent a categorical feature, such as the species of conifers. MLJ mitigates such ambiguity by: (i) distinguishing between the machine and scientific type of scalar data; (ii) disallowing the representation of multiple scientific types by the same machine type during learning; and (iii) establising a convention for what scientific types a given machine type may represent (see the table at the end of Getting Started).

Explicitly specifying scientific types during the construction of a MLJ task is the user's opportunity to articulate how the supplied data should be interpreted.

Learning objectives

In MLJ specifying a learing objective means specifying: (i) whether learning is supervised or not; (ii) whether, in the supervised case, predictions are to be probabilistic or deterministic; and (iii) what part of the data is relevant and what role is each part to play.

Sample usage

Load a built-in task:

using MLJ
task = load_iris()

SupervisedTask @ 3…93

Extract input and target:

X, y = task()
X[1:3, :]

sepal_length	sepal_width	petal_length	petal_width

Float64	Float64	Float64	Float64

3 rows × 4 columns

1	5.1	3.5	1.4	0.2

2	4.9	3.0	1.4	0.2

3	4.7	3.2	1.3	0.2

Now, starting with some tabular data...

using RDatasets
df = dataset("boot", "channing");
first(df, 4)

Sex	Entry	Exit	Time	Cens

Categorical…	Int32	Int32	Int32	Int32

4 rows × 5 columns

1	Male	782	909	127	1

2	Male	1020	1128	108	1

3	Male	856	969	113	1

4	Male	915	957	42	1

...we can check MLJ's interpretation of that data:

scitypes(df)

(Sex = Multiclass{2},
 Entry = Count,
 Exit = Count,
 Time = Count,
 Cens = Count,)

And construct a task by wrapping the data in a learning objective, and coercing the data into a form MLJ will correctly interpret. (The middle three fields of df refer to ages, in months, the last is a flag.):

task = supervised(data=df,
                  target=:Exit,
                  ignore=:Time,
                  is_probabilistic=true,
                  types=Dict(:Entry=>Continuous,
                             :Exit=>Continuous,
                             :Cens=>Multiclass))
first(task.X, 4)

Shuffle the rows of a task:

task = load_iris()
using Random
rng = MersenneTwister(1234)
shuffle!(rng, task) # rng is optional
task[1:4].y

4-element CategoricalArrays.CategoricalArray{String,1,UInt32}:
 "virginica"
 "virginica"
 "setosa"   
 "virginica"

Counting and selecting rows of a task:

nrows(task)

150

task[1:2].y

2-element CategoricalArrays.CategoricalArray{String,1,UInt32}:
 "virginica"
 "virginica"

Listing the models available to complete a task:

models(task)

Dict{String,Any} with 2 entries:
  "MLJ"         => Any["DeterministicConstantClassifier"]
  "ScikitLearn" => Any["SVMLClassifier", "SVMNuClassifier", "SVMClassifier"]

Binding a model to a task and evalutating performance:

@load DecisionTreeClassifier
mach = machine(DecisionTreeClassifier(), task)
evaluate!(mach, operation=predict_mode, resampling=Holdout(), measure=misclassification_rate, verbosity=0)

0.08888888888888889

API Reference

MLJ.supervised — Function.

task = supervised(data=nothing, 
                  types=Dict(), 
                  target=nothing,  
                  ignore=Symbol[], 
                  is_probabilistic=false, 
                  verbosity=1)

Construct a supervised learning task with input features X and target y, where: y is the column vector from data named target, if this is a single symbol, or, a vector of tuples, if target is a vector; X consists of all remaining columns of data not named in ignore, and is a table unless it has only one column, in which case it is a vector.

The data types of elements in a column of data named as a key of the dictionary types are coerced to have a scientific type given by the corresponding value. Possible values are Continuous, Multiclass, OrderedFactor and Count. So, for example, types=Dict(:x1=>Count) means elements of the column of data named :x1 will be coerced into integers (whose scitypes are always Count).

task = supervised(X, y; 
                  input_is_multivariate=true, 
                  is_probabilistic=false, 
                  verbosity=1)

A more customizable constructor, this returns a supervised learning task with input features X and target y, where: X must be a table or vector, according to whether it is multivariate or univariate, while y must be a vector whose elements are scalars, or tuples scalars (of constant length for ordinary multivariate predictions, and of variable length for sequence prediction). Table rows must correspond to patterns and columns to features. Type coercion is not available for this constructor (but see also coerce).

X, y = task()

Returns the input X and target y of the task, also available as task.X and task.y.

source

MLJ.unsupervised — Function.

task = unsupervised(data=nothing, types=Dict(), ignore=Symbol[], verbosity=1)

Construct an unsupervised learning task with given input data, which should be a table or, in the case of univariate inputs, a single vector.

The data types of elements in a column of data named as a key of the dictionary types are coerced to have a scientific type given by the corresponding value. Possible values are Continuous, Multiclass, OrderedFactor and Count. So, for example, types=Dict(:x1=>Count) means elements of the column of data named :x1 will be coerced into integers (whose scitypes are always Count).

Rows of data must correspond to patterns and columns to features. Columns in data whose names appear in ignore are ignored.

X = task()

Return the input data in form to be used in models.

See also scitype, scitype_union, scitypes

source

MLJ.models — Method.

models(; show_dotted=false)

List all models as a dictionary indexed on package name. Models available for immediate use appear under the key "MLJ".

By declaring show_dotted=true models not in the top-level of the current namespace - which require dots to call, such as MLJ.DeterministicConstantModel - are also included.

models(task; show_dotted=false)

List all models matching the specified task.

See also: localmodels

source

MLJ.localmodels — Method.

localmodels()

List all models available for immediate use. Equivalent to models()["MLJ"]

localmodels(task)

List all such models additionally matching the specified task. Equivalent to models(task)["MLJ"].

source