Evaluating Model Performance

Evaluating Model Performance

MLJ allows quick evaluation of a supervised model's performance against a battery of selected losses or scores. For more on available performance measures, see Performance Measures.

In addition to hold-out and cross-validation, the user can specify their own list of train/test pairs of row indices for resampling, or define their own re-usable resampling strategies.

For simultaneously evaluating multiple models and/or data sets, see Benchmarking.

Evaluating against a single measure

julia> using MLJ

julia> X = (a=rand(12), b=rand(12), c=rand(12));

julia> y = X.a + 2X.b + 0.05*rand(12);

julia> model = @load RidgeRegressor pkg=MultivariateStats
MLJModels.MultivariateStats_.RidgeRegressor(lambda = 1.0,) @ 9…14

julia> cv=CV(nfolds=3)
CV(nfolds = 3,
   shuffle = false,
   rng = MersenneTwister(UInt32[0xb7dcb697, 0x619e3519, 0xdb6967a7, 0xa827f1f4]) @ 417,) @ 1…85

julia> evaluate(model, X, y, resampling=cv, measure=l2, verbosity=0)
(measure = MLJBase.L2[l2],
 measurement = [0.2467251654691582],
 per_fold = Array{Float64,1}[[0.20333584169253877, 0.3220490345360833, 0.21479062017885253]],
 per_observation = Array{Array{Float64,1},1}[[[0.016262220909036403, 0.15417897689343904, 0.3192852173703758, 0.3236169515973038], [0.05621613600208008, 0.13776114727084526, 0.7667949221564091, 0.32742393271499876], [0.41545132123878953, 0.024500894971138137, 0.11474954729723535, 0.30446071720824713]]],)

Alternatively, instead of applying evaluate to a model + data, one may call evaluate! on an existing machine wrapping the model in data:

julia> mach = machine(model, X, y)
Machine{RidgeRegressor} @ 9…06

julia> evaluate!(mach, resampling=cv, measure=l2, verbosity=0)
(measure = MLJBase.L2[l2],
 measurement = [0.2467251654691582],
 per_fold = Array{Float64,1}[[0.20333584169253877, 0.3220490345360833, 0.21479062017885253]],
 per_observation = Array{Array{Float64,1},1}[[[0.016262220909036403, 0.15417897689343904, 0.3192852173703758, 0.3236169515973038], [0.05621613600208008, 0.13776114727084526, 0.7667949221564091, 0.32742393271499876], [0.41545132123878953, 0.024500894971138137, 0.11474954729723535, 0.30446071720824713]]],)

(The latter call is a mutating call as the learned parameters stored in the machine potentially change. )

Multiple measures

julia> evaluate!(mach,
                 resampling=cv,
                 measure=[l1, rms, rmslp1], verbosity=0)
(measure = MLJBase.Measure[l1, rms, rmslp1],
 measurement = [0.45015458818996307, 0.49671437010535363, 0.23097441409568137],
 per_fold = Array{Float64,1}[[0.4135266037103326, 0.5140348187813344, 0.4229023420782221], [0.45092775662243145, 0.5674936427274612, 0.4634550897108074], [0.28235151835593014, 0.21275556901741158, 0.1872437658099947]],
 per_observation = Union{Missing, Array{Array{Float64,1},1}}[Array{Float64,1}[[0.12752341317984084, 0.3926563088674866, 0.565053287195443, 0.56887340559856], [0.23709942218841462, 0.37116188822513185, 0.8756682717538697, 0.5722096929579215], [0.6445551343669442, 0.15652761727930997, 0.3387470255179156, 0.5517795911487187]], missing, missing],)

Custom measures and weighted measures

julia> my_loss(yhat, y) = maximum((yhat - y).^2);

julia> my_per_observation_loss(yhat, y) = abs.(yhat - y);

julia> MLJ.reports_each_observation(::typeof(my_per_observation_loss)) = true;

julia> my_weighted_score(yhat, y) = 1/mean(abs.(yhat - y));

julia> my_weighted_score(yhat, y, w) = 1/mean(abs.((yhat - y).^w));

julia> MLJ.supports_weights(::typeof(my_weighted_score)) = true;

julia> MLJ.orientation(::typeof(my_weighted_score)) = :score;

julia> holdout = Holdout(fraction_train=0.8)
Holdout(fraction_train = 0.8,
        shuffle = false,
        rng = MersenneTwister(UInt32[0xb7dcb697, 0x619e3519, 0xdb6967a7, 0xa827f1f4]) @ 417,) @ 1…34

julia> weights = [1, 1, 2, 1, 1, 2, 3, 1, 1, 2, 3, 1];

julia> evaluate!(mach,
                 resampling=CV(nfolds=3),
                 measure=[my_loss, my_per_observation_loss, my_weighted_score, l1],
                 weights=weights, verbosity=0)
┌ Warning: Sample weights ignored in evaluations of the following measures, as unsupported: 
│ my_loss, my_per_observation_loss 
└ @ MLJ ~/build/alan-turing-institute/MLJ.jl/src/resampling.jl:420
(measure = Any[Main.ex-evaluation_of_supervised_models.my_loss, Main.ex-evaluation_of_supervised_models.my_per_observation_loss, Main.ex-evaluation_of_supervised_models.my_weighted_score, l1],
 measurement = [0.5019543983308342, 0.45015458818996307, 2.8289833196815, 0.4671948726827487],
 per_fold = Array{Float64,1}[[0.3236169515973038, 0.7667949221564091, 0.41545132123878953], [0.4135266037103326, 0.5140348187813344, 0.4229023420782221], [2.84022657918457, 2.4713810266682072, 3.175342353191722], [0.4438319404073547, 0.5969482438368869, 0.3608044338040043]],
 per_observation = Union{Missing, Array{Array{Float64,1},1}}[missing, Array{Float64,1}[[0.12752341317984084, 0.3926563088674866, 0.565053287195443, 0.56887340559856], [0.23709942218841462, 0.37116188822513185, 0.8756682717538697, 0.5722096929579215], [0.6445551343669442, 0.15652761727930997, 0.3387470255179156, 0.5517795911487187]], missing, Array{Float64,1}[[0.10201873054387267, 0.31412504709398925, 0.9040852595127088, 0.455098724478848], [0.1354853841076655, 0.4241850151144364, 1.5011456087209196, 0.32697696740452653], [0.36831721963825387, 0.17888870546206853, 0.5807091866021411, 0.31530262351355354]]],)

User-specified train/test sets

Users can either provide their own list of train/test pairs of row indices for resampling, as in this example:

julia> fold1 = 1:6; fold2 = 7:12;

julia> evaluate!(mach,
                 resampling = [(fold1, fold2), (fold2, fold1)],
                 measure=[l1, l2], verbosity=0)
(measure = MLJBase.Measure[l1, l2],
 measurement = [0.9541966897275136, 1.1567590156725103],
 per_fold = Array{Float64,1}[[1.196033905786677, 0.7123594736683504], [1.7248697310886234, 0.5886483002563972]],
 per_observation = Array{Array{Float64,1},1}[[[2.017812551805672, 1.1834853731743513, 1.427662387216834, 1.0276733576055, 0.19936775800701256, 1.320202006910691], [0.10048277183572552, 0.7562759270622603, 0.9571460472668574, 0.8989446531279925, 0.7224762956654496, 0.8388311470518165]], [[4.071567494224517, 1.4006376285176336, 2.038219891873669, 1.0561125299321616, 0.03974750293274272, 1.7429333390510162], [0.010096787435790473, 0.5719532778538813, 0.9161285557985691, 0.8081014893874068, 0.5219719977984701, 0.7036376932642662]]],)

Or define their own re-usable ResamplingStrategy objects, - see Custom resampling strategies below.

Built-in resampling strategies

MLJ.Holdout — Type.

holdout = Holdout(; fraction_train=0.7,
                     shuffle=false,
                     rng=Random.GLOBAL_RNG)

Holdout resampling strategy, for use in evaluate!, evaluate and in tuning.

train_test_pairs(holdout, rows)

Returns the pair [(train, test)], where train and test are vectors such that rows=vcat(train, test) and length(train)/length(test) ≈ fraction_train.

If rng is an integer, then MersenneTwister(rng) is the random number generator used for shuffling rows. Otherwise some AbstractRNG object is expected.

source

MLJ.CV — Type.

cv = CV(; nfolds=6,  shuffle=false, rng=Random.GLOBAL_RNG)

Cross-validation resampling strategy, for use in evaluate!, evaluate and tuning.

train_test_pairs(cv, rows)

Returns an nfolds-length iterator of (train, test) pairs of vectors (row indices), where each train and test is a sub-vector of rows. The test vectors are mutually exclusive and exhaust rows. Each train vector is the complement of the corresponding test vector. With no shuffling, the order of rows is preserved, in the sense that rows coincides precisely with the concatenation of the test vectors, in the order they are generated. All but the last test vector have equal length.

Declaring shuffle=true results in rows being shuffled first. If rng is an integer, then MersenneTwister(rng) is the random number generator used for shuffling rows. Otherwise some AbstractRNG object is expected.

source

MLJ.StratifiedCV — Type.

stratified_cv = StratifiedCV(; nfolds=6,  
                               shuffle=false, 
                               rng=Random.GLOBAL_RNG)

Stratified cross-validation resampling strategy, for use in evaluate!, evaluate and in tuning. Applies only to classification problems (OrderedFactor or Multiclass targets).

train_test_pairs(stratified_cv, rows, y)

Returns an nfolds-length iterator of (train, test) pairs of vectors (row indices) where each train and test is a sub-vector of rows. The test vectors are mutually exclusive and exhaust rows. Each train vector is the complement of the corresponding test vector.

Unlike regular cross-validation, the distribution of the levels of the target y corresponding to each train and test is constrained, as far as possible, to replicate that of y[rows] as a whole.

Specifically, the data is split into a number of groups on which y is constant, and each individual group is resampled according to the ordinary cross-validation strategy CV(nfolds=nfolds). To obtain the final (train, test) pairs of row indices, the per-group pairs are collated in such a way that each collated train and test respects the original order of rows (after shuffling, if shuffle=true).

If rng is an integer, then MersenneTwister(rng) is the random number generator used for shuffling rows. Otherwise some AbstractRNG object is expected.

source

Custom resampling strategies

To define your own resampling strategy, make relevant parameters of your strategy the fields of a new type MyResamplingStrategy <: MLJ.ResamplingStrategy, and implement one of the following methods:

MLJ.train_test_pairs(my_strategy::MyResamplingStrategy, rows)
MLJ.train_test_pairs(my_strategy::MyResamplingStrategy, rows, y)
MLJ.train_test_pairs(my_strategy::MyResamplingStrategy, rows, X, y)

Each method takes a vector of indices rows and return a vector [(t1, e1), (t2, e2), ... (tk, ek)] of train/test pairs of row indices selected from rows. Here X, y are the input and target data (ignored in simple strategies, such as Holdout and CV).

Here is the code for the Holdout strategy as an example:

struct Holdout <: MLJ.ResamplingStrategy
    fraction_train::Float64
    shuffle::Bool
    rng::Union{Int,AbstractRNG}
	
    function Holdout(fraction_train, shuffle, rng)
        0 < fraction_train < 1 || 
		error("`fraction_train` must be between 0 and 1.")
        return new(fraction_train, shuffle, rng)
    end
end

# Keyword Constructor
function Holdout(; fraction_train::Float64=0.7,
                   shuffle::Bool=false,
                   rng::Union{Int,AbstractRNG}=Random.GLOBAL_RNG)
    Holdout(fraction_train, shuffle, rng)
end

function MLJ.train_test_pairs(holdout::Holdout, rows)
    if holdout.rng isa Integer
        rng = MersenneTwister(holdout.rng)
    else
        rng = holdout.rng
    end
    train, evalu = partition(rows, holdout.fraction_train,
                             shuffle=holdout.shuffle, rng=rng)
    return [(train, evalu),]
end

API

MLJ.evaluate! — Function.

evaluate!(mach,
          resampling=CV(),
          measure=nothing,
          weights=nothing,
          operation=predict,
          acceleration=DEFAULT_RESOURCE[],
          force=false,
          verbosity=1)

Estimate the performance of a machine mach wrapping a supervised model in data, using the specified resampling strategy (defaulting to 6-fold cross-validation) and measure, which can be a single measure or vector.

Do subtypes(MLJ.ResamplingStrategy) to obtain a list of available resampling strategies. If resampling is not an object of type MLJ.ResamplingStrategy, then a vector of pairs (of the form (train_rows, test_rows) is expected. For example, setting

resampling = [(1:100), (101:200)),
               (101:200), (1:100)]

gives two-fold cross-validation using the first 200 rows of data.

If resampling isa MLJ.ResamplingStrategy then one may optionally restrict the data used in evaluation by specifying rows.

An optional weights vector may be passed for measures that support sample weights (MLJ.supports_weights(measure) == true), which is ignored by those that don't.

Important: If mach already wraps sample weights w (as in mach = machine(model, X, y, w)) then these weights, which are used for training, are automatically passed to the measures for evaluation. However, for evaluation purposes, any weights specified as a keyword argument will take precedence over w.

User-defined measures are supported; see the manual for details.

If no measure is specified, then default_measure(mach.model) is used, unless this default is nothing and an error is thrown.

The acceleration keyword argument is used to specify the compute resource (a subtype of ComputationalResources.AbstractResource) that will be used to accelerate/parallelize the resampling operation.

Although evaluate! is mutating, mach.model and mach.args are untouched.

source

MLJBase.evaluate — Function.

evaluate(model, X, y; measure=nothing, options...)
evaluate(model, X, y, w; measure=nothing, options...)

Evaluate the performance of a supervised model model on input data X and target y, optionally specifying sample weights w for training, where supported. The same weights are passed to measures that support sample weights, unless this behaviour is overridden by explicitly specifying the option weights=....

See the machine version evaluate! for the complete list of options.

source