Evaluating Model Performance

Evaluating Model Performance

MLJ allows quick evaluation of a supervised model's performance against a battery of selected losses or scores. For more on available performance measures, see Performance Measures.

In addition to hold-out and cross-validation, the user can specify their own list of train/test pairs of row indices for resampling, or define their own re-usable resampling strategies.

For simultaneously evaluating multiple models and/or data sets, see Benchmarking.

Evaluating against a single measure

julia> using MLJ

julia> X = (a=rand(12), b=rand(12), c=rand(12));

julia> y = X.a + 2X.b + 0.05*rand(12);

julia> model = @load RidgeRegressor pkg=MultivariateStats
MLJModels.MultivariateStats_.RidgeRegressor(lambda = 0.0,) @ 9…01

julia> cv=CV(nfolds=3)
CV(nfolds = 3,
   shuffle = false,
   rng = MersenneTwister(UInt32[0x000004d2]),) @ 1…22

julia> evaluate(model, X, y, resampling=cv, measure=l2, verbosity=0)
(measure = MLJBase.L2[l2],
 measurement = [0.00019423054640673845],
 per_fold = Array{Float64,1}[[0.00029911633535035294, 1.2093864400094128e-5, 0.00027148143946976837]],
 per_observation = Array{Array{Float64,1},1}[[[0.00019962839967756598, 4.610201818475497e-5, 0.00047872500151901906, 0.00047200992202007185], [7.98802321816122e-6, 3.1472844414552503e-6, 1.7005856609009607e-5, 2.0234293331750436e-5], [1.7535171639283048e-6, 0.0006484306350234522, 0.0004336570483019112, 2.0845573897817193e-6]]],)

Alternatively, instead of applying evaluate to a model + data, one may call evaluate! on an existing machine wrapping the model in data:

julia> mach = machine(model, X, y)
Machine{RidgeRegressor} @ 1…58

julia> evaluate!(mach, resampling=cv, measure=l2, verbosity=0)
(measure = MLJBase.L2[l2],
 measurement = [0.00019423054640673845],
 per_fold = Array{Float64,1}[[0.00029911633535035294, 1.2093864400094128e-5, 0.00027148143946976837]],
 per_observation = Array{Array{Float64,1},1}[[[0.00019962839967756598, 4.610201818475497e-5, 0.00047872500151901906, 0.00047200992202007185], [7.98802321816122e-6, 3.1472844414552503e-6, 1.7005856609009607e-5, 2.0234293331750436e-5], [1.7535171639283048e-6, 0.0006484306350234522, 0.0004336570483019112, 2.0845573897817193e-6]]],)

(The latter call is a mutating call as the learned parameters stored in the machine potentially change. )

Multiple measures

julia> evaluate!(mach,
                 resampling=cv,
                 measure=[l1, rms, rmslp1], verbosity=0)
(measure = MLJBase.Measure[l1, rms, rmslp1],
 measurement = [0.01056696582815663, 0.01241643242435523, 0.005127487784178825],
 per_fold = Array{Float64,1}[[0.01613110317850433, 0.003305609530642406, 0.01226418477532315], [0.017294980062155404, 0.0034776233838778644, 0.016476693827032422], [0.0072899845069067035, 0.0014073240927141794, 0.006685154752915592]],
 per_observation = Union{Missing, Array{Array{Float64,1},1}}[Array{Float64,1}[[0.014128991460028772, 0.006789846698177726, 0.021879785225614512, 0.02172578933019631], [0.002826309115818937, 0.0017740587480281622, 0.004123815782622886, 0.004498254476099639], [0.001324204351272229, 0.02546430118859444, 0.02082443392512534, 0.0014437996363005912]], missing, missing],)

Custom measures and weighted measures

julia> my_loss(yhat, y) = maximum((yhat - y).^2);

julia> my_per_observation_loss(yhat, y) = abs.(yhat - y);

julia> MLJ.reports_each_observation(::typeof(my_per_observation_loss)) = true;

julia> my_weighted_score(yhat, y) = 1/mean(abs.(yhat - y));

julia> my_weighted_score(yhat, y, w) = 1/mean(abs.((yhat - y).^w));

julia> MLJ.supports_weights(::typeof(my_weighted_score)) = true;

julia> MLJ.orientation(::typeof(my_weighted_score)) = :score;

julia> holdout = Holdout(fraction_train=0.8)
Holdout(fraction_train = 0.8,
        shuffle = false,
        rng = MersenneTwister(UInt32[0x000004d2]),) @ 1…42

julia> weights = [1, 1, 2, 1, 1, 2, 3, 1, 1, 2, 3, 1];

julia> evaluate!(mach,
                 resampling=CV(nfolds=3),
                 measure=[my_loss, my_per_observation_loss, my_weighted_score, l1],
                 weights=weights, verbosity=0)
┌ Warning: weights ignored in evaluations of the following measures, as unsupported: 
│ my_loss, my_per_observation_loss 
└ @ MLJ ~/build/alan-turing-institute/MLJ.jl/src/resampling.jl:265
(measure = Any[Main.ex-evaluation_of_supervised_models.my_loss, Main.ex-evaluation_of_supervised_models.my_per_observation_loss, Main.ex-evaluation_of_supervised_models.my_weighted_score, l1],
 measurement = [0.00038246330995807395, 0.01056696582815663, 602.116528307066, 0.012399043509117422],
 per_fold = Array{Float64,1}[[0.00047872500151901906, 2.0234293331750436e-5, 0.0006484306350234522], [0.01613110317850433, 0.003305609530642406, 0.01226418477532315], [92.75716680271606, 545.8678414465824, 1167.7245766718995], [0.017280839587926367, 0.00332058977654908, 0.016595701162876817]],
 per_observation = Union{Missing, Array{Array{Float64,1},1}}[missing, Array{Float64,1}[[0.014128991460028772, 0.006789846698177726, 0.021879785225614512, 0.02172578933019631], [0.002826309115818937, 0.0017740587480281622, 0.004123815782622886, 0.004498254476099639], [0.001324204351272229, 0.02546430118859444, 0.02082443392512534, 0.0014437996363005912]], missing, Array{Float64,1}[[0.011303193168023019, 0.005431877358542181, 0.03500765636098322, 0.017380631464157047], [0.001615033780467964, 0.002027495712032185, 0.007069398484496376, 0.002570431129199794], [0.000756688200726988, 0.029102058501250787, 0.03569902958592915, 0.0008250283636003378]]],)

User-specified train/test sets

Users can either provide their own list of train/test pairs of row indices for resampling, as in this example:

julia> fold1 = 1:6; fold2 = 7:12;

julia> evaluate!(mach,
                 resampling = [(fold1, fold2), (fold2, fold1)],
                 measure=[l1, l2], verbosity=0)
(measure = MLJBase.Measure[l1, l2],
 measurement = [0.013981072803757238, 0.00028834247102562586],
 per_fold = Array{Float64,1}[[0.015682333461501535, 0.012279812146012944], [0.0003570312868826416, 0.0002196536551686101]],
 per_observation = Array{Array{Float64,1},1}[[[0.017998786096304453, 0.007868701060771865, 0.00833836405147892, 0.03406594532546947, 0.022732122996760307, 0.003090081238224185], [0.01720331347035875, 0.006264321779738902, 0.021627105561553783, 0.021969960361953422, 0.0010221654630355737, 0.0055920062394372305]], [[0.0003239563009405225, 6.191645638379228e-5, 6.952831505499594e-5, 0.0011604886309178753, 0.0005167494159398388, 9.548602058825113e-6], [0.00029595399435942683, 3.924172736011116e-5, 0.0004677316949705906, 0.00048267915830580457, 1.0448222338227286e-6, 3.127053378190492e-5]]],)

Or define their own re-usable ResamplingStrategy objects, - see Custom resampling strategies below.

Resampling strategies

Holdout and CV (cross-validation) resampling strategies are available:

MLJ.Holdout — Type.

Holdout(; fraction_train=0.7,
          shuffle=false,
          rng=Random.GLOBAL_RNG)

Single train-test split with a (randomly selected) portion of the data being selected for training and the rest for testing.

If rng is an integer, then MersenneTwister(rng) is the random number generator used for shuffling rows. Otherwise some AbstractRNG object is expected.

source

MLJ.CV — Type.

CV(; nfolds=6,  shuffle=false, rng=Random.GLOBAL_RNG)

Cross validation resampling where the data is (randomly) partitioned in nfolds folds and the model is evaluated nfolds times, each time taking one fold for testing and the remaining folds for training.

For instance, if nfolds=3 then the data will be partitioned in three folds A, B and C and the model will be trained three times, first with A and B and tested on C, then on A, C and tested on B and finally on B, C and tested on A.

If rng is an integer, then MersenneTwister(rng) is the random number generator used for shuffling rows. Otherwise some AbstractRNG object is expected.

source

Custom resampling strategies

To define your own resampling strategy, make relevant parameters of your strategy the fields of a new type MyResamplingStrategy <: MLJ.ResamplingStrategy, and implement MLJ.train_test_pairs(my_strategy::MyStragegy, rows), a method which will take a vector of indices rows and return a vector [(t1, e1), (t2, e2), ... (tk, ek)] of train/test pairs of row indices selected from rows. Here is the code for the Holdout strategy as an example:

struct Holdout <: ResamplingStrategy
    fraction_train::Float64
    shuffle::Bool
    rng::Union{Int,AbstractRNG}
	
    function Holdout(fraction_train, shuffle, rng)
        0 < fraction_train < 1 || 
		error("`fraction_train` must be between 0 and 1.")
        return new(fraction_train, shuffle, rng)
    end
end

# Keyword Constructor
function Holdout(; fraction_train::Float64=0.7,
                   shuffle::Bool=false,
                   rng::Union{Int,AbstractRNG}=Random.GLOBAL_RNG)
    Holdout(fraction_train, shuffle, rng)
end

function train_test_pairs(holdout::Holdout, rows)
    if holdout.rng isa Integer
        rng = MersenneTwister(holdout.rng)
    else
        rng = holdout.rng
    end
    train, evalu = partition(rows, holdout.fraction_train,
                             shuffle=holdout.shuffle, rng=rng)
    return [(train, evalu),]
end

API

MLJ.evaluate! — Function.

evaluate!(mach,    
          resampling=CV(), 
          measure=nothing, 
          weights=nothing,
          operation=predict,  
          parallel=true,
          force=false, 
          verbosity=1)

Estimate the performance of a machine mach wrapping a supervised model in data, using the specified resampling strategy (defaulting to 6-fold cross-validation) and measure, which can be a single measure or vector.

Do subtypes(MLJ.ResamplingStrategy) to obtain a list of available resampling strategies. If resampling is not an object of type MLJ.ResamplingStrategy, then a vector of pairs (of the form (train_rows, test_rows) is expected. For example, setting

resampling = [(1:100), (101:200)), 
               (101:200), (1:100)]

gives two-fold cross-validation using the first 200 rows of data.

If resampling isa MLJ.ResamplingStrategy then one may optionally restrict the data used in evaluation by specifying rows.

An optional weights vector may be passed for measures that support sample weights (MLJ.supports_weights(measure) == true), which is ignored by those that don't.

User-defined measures are supported; see the manual for details.

If no measure is specified, then default_measure(mach.model) is used, unless this default is nothing and an error is thrown.

Although evaluate! is mutating, mach.model and mach.args are untouched.

source

MLJBase.evaluate — Function.

evaluate(model, X, y; measure=nothing, options...)

Evaluate the performance of a supervised model model on input data X and target y. See the machine version evaluate! for options.

source