OpenML Integration

OpenML provides an integration platform for carrying out and comparing machine learning solutions across a broad collection of public datasets and software platforms. Integration of MLJ with OpenML is a work in progress.

Loading IRIS Dataset

As an example, we will try to load the iris dataset using OpenML.load(taskID).

using MLJ

Task ID

OpenML.load requires task ID of the the dataset to be loaded. This ID can be found on the OpenML website. The task ID for the iris dataset is 61, as mentioned in this OpenML Page

julia> rowtable = OpenML.load(61)
Error occurred : HTTP.ExceptionRequest.StatusError(412, "GET", "/api/v1/json/data/61", HTTP.Messages.Response:
"""
HTTP/1.1 412 Precondition Failed
Date: Wed, 07 Apr 2021 06:14:09 GMT
Server: Apache/2.4.29 (Ubuntu)
Set-Cookie: ci_session=ivdr08ut2s6etf9u69idpi34nr2qqaki; expires=Wed, 07-Apr-2021 08:14:09 GMT; Max-Age=7200; path=/; HttpOnly
Expires: Thu, 19 Nov 1981 08:52:00 GMT
Cache-Control: no-store, no-cache, must-revalidate
Pragma: no-cache
Content-Length: 219
Content-Type: text/xml; charset=utf-8

<oml:error xmlns:oml="http://openml.org/openml">
	<oml:code>107</oml:code>
	<oml:message>Database connection error. Usually due to high server load. Please wait for N seconds and try again. </oml:message>
	</oml:error>
""")
ERROR: MethodError: no method matching getindex(::Nothing, ::String)

Converting to DataFrame

julia> using DataFrames

julia> df = DataFrame(rowtable)
ERROR: UndefVarError: rowtable not defined

julia> df2 = coerce(df, :class=>Multiclass)
ERROR: UndefVarError: df not defined

MLJOpenML.load — Function

MLJOpenML.load(id)

Load the OpenML dataset with specified id, from those listed on the OpenML site.

Returns a "row table", i.e., a Vector of identically typed NamedTuples. A row table is compatible with the Tables.jl interface and can therefore be readily converted to other compatible formats. For example:

using DataFrames
rowtable = MLJOpenML.load(61);
df = DataFrame(rowtable);

using MLJ
df2 = coerce(df, :class=>Multiclass)