Working with Tasks

Working with Tasks

In MLJ a task is a synthesis of three elements: data, an interpretation of data, and a learning objective. Once one has a task - no more and no less - one is ready to choose a learning model.

Scientific types and the interpretation of data

Generally the columns of a table, such as a DataFrame, represents real quantities. However, the nature of a quantity is not always clear from the representation. For example, we might count phone calls using the UInt32 type but also use UInt32 to represent a categorical feature, such as the species of conifers. MLJ mitigates such ambiguity by: (i) distinguishing between the machine and scientific type of scalar data; (ii) disallowing the representation of multiple scientific types by the same machine type during learning; and (iii) establising a convention for what scientific types a given machine type may represent (see the table at the end of Getting Started).

Explicitly specifying scientific types during the construction of a MLJ task is the user's opportunity to articulate how the supplied data should be interpreted.

WIP: At present scitypes cannot be specified and the user must manually coerce data before task construction.

Learning objectives

In MLJ specifying a learing objective means specifying: (i) whether learning is supervised or not; (ii) whether, in the supervised case, predictions are to be probabilistic or deterministic; and (iii) what part of the data is relevant and what role is each part to play.

Sample usage

# load a built-in task:
using MLJ
task = load_iris()

[0m[1mSupervisedTask @ 1…26[22m

# deconstruct:
X, y = task()
X[1:3, :]

sepal_length	sepal_width	petal_length	petal_width

Float64	Float64	Float64	Float64

3 rows × 4 columns

1	5.1	3.5	1.4	0.2

2	4.9	3.0	1.4	0.2

3	4.7	3.2	1.3	0.2

# reconstruct:
df = copy(X)
df.species = y
task = SupervisedTask(data=df, target=:species, is_probabilistic=true)
show(task, 1)

# [0m[1mSupervisedTask @ 1…20[22m:
X                       =>   (omitted DataFrames.DataFrame)
y                       =>   (omitted CategoricalArrays.CategoricalArray{String,1,UInt32,String,CategoricalArrays.CategoricalString{UInt32},Union{}} of size (150,))
is_probabilistic        =>   true
target_scitype          =>   Multiclass{3}
input_scitypes          =>   Continuous
input_is_multivariate   =>   true

models(task)

Dict{String,Any} with 4 entries:
  "MLJ"          => Any["ConstantClassifier"]
  "DecisionTree" => Any["DecisionTreeClassifier"]
  "NaiveBayes"   => Any["GaussianNBClassifier"]
  "XGBoost"      => Any["XGBoostClassifier"]

@load DecisionTreeClassifier
mach = machine(DecisionTreeClassifier(target_type=String), task)
evaluate!(mach, operation=predict_mode, resampling=Holdout(), measure=misclassification_rate, verbosity=0)

API

MLJBase.UnsupervisedTask — Type.

task = UnsupervisedTask(data=nothing, ignore=Symbol[], input_is_multivariate=true)

Construct an unsupervised learning task with given input data, which should be a table or, in the case of univariate inputs, a single vector.

Rows of data must correspond to patterns and columns to features. Columns in data whose names appear in ignore are ignored.

X = task()

Return the input data in form to be used in models.

source

MLJBase.SupervisedTask — Type.

task = SupervisedTask(X, y; is_probabilistic=nothing, input_is_multivariate=true, target_is_multivariate=false)

Construct a supervised learning task with input features X and target y. Both X and y must be tables or vectors, according to whether they are multivariate or univariate. Table rows must correspond to patterns and columns to features. The boolean keyword argument is_probabilistic must be specified.

task = SupervisedTask(data=nothing, is_probabilistic=nothing, target=nothing, ignore=Symbol[], input_is_multivariate=true)

Construct a supervised learning task with input features X and target y, where y is the column vector from data named target (if this is a single symbol) or, a table whose columns are those named in target (if this is vector); X consists of all remaining columns of data not named in ignore.

X, y = task()

Returns the input X and target y of the task, also available as task.X and task.y.

source

MLJ.models — Method.

models(; show_dotted=false)

List all models as a dictionary indexed on package name. Models available for immediate use appear under the key "MLJ".

By declaring show_dotted=true models not in the top-level of the current namespace - which require dots to call, such as MLJ.DeterministicConstantModel - are also included.

models(task; show_dotted=false)

List all models matching the specified task.

See also: localmodels

source

MLJ.localmodels — Method.

localmodels()

List all models available for immediate use. Equivalent to models()["MLJ"]

localmodels(task)

List all such models additionally matching the specified task. Equivalent to models(task)["MLJ"].

source